fix: use deep comparison in useSyncExternalStore selector
If the user is trying to select an object from the store instead of a
primitive, the reference of the returned value will keep changing every
time, resulting in infinite re-rendering.

Adding a deep comparison check here with the previously selected state
will prevent this from happening. This might potentially cause a large
performance degradation, so will need to be benchmarked